@@ -9,14 +9,27 @@
     "eject": "react-scripts eject"
   },
   "dependencies": {
-    "react": "^18.0.0",
+    "@types/react-router-dom": "^5.3.3",
+    "react": "^18.3.1",
     "react-dom": "^18.0.0",
+    "react-router-dom": "^7.9.1",
     "react-scripts": "5.0.0",
     "typescript": "^4.0.0"
   },
   "devDependencies": {
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0"
+    "@types/babel__traverse": "^7.28.0",
+    "@types/eslint": "^9.6.1",
+    "@types/estree": "^1.0.8",
+    "@types/html-minifier-terser": "^7.0.2",
+    "@types/istanbul-lib-coverage": "^2.0.6",
+    "@types/istanbul-lib-report": "^3.0.3",
+    "@types/istanbul-reports": "^3.0.4",
+    "@types/node": "^24.4.0",
+    "@types/prettier": "^3.0.0",
+    "@types/react": "^18.3.24",
+    "@types/react-dom": "^18.0.0",
+    "@types/semver": "^7.7.1",
+    "@types/stack-utils": "^2.0.3"
   },
   "eslintConfig": {
     "extends": [
