@@ -1,5 +1,5 @@
 import React from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import MobileNavbar from './components/MobileNavbar';
 import Home from './pages/Home';
 import './styles/mobile.css';
@@ -9,9 +9,9 @@ const App: React.FC = () => {
     <Router>
       <div>
         <MobileNavbar />
-        <Switch>
-          <Route path="/" exact component={Home} />
-        </Switch>
+        <Routes>
+          <Route path="/" element={<Home />} />
+        </Routes>
       </div>
     </Router>
   );
